requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
culqipy==0.2.6
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.4.0
django-filter==2.3.0
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
idna==2.10
Pillow==7.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.5
pycparser==2.20
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
